refactor: Moved code of systems and components from Main to separate files + modified gravity system
Main.cpp now includes GameSpecificComponents.h and GameSpecificSystems.h
- GameSpecificComponents.h:
	Contains definitions of constants and components with no behaviour
- GameSpecificSystems.h:
	Includes .h files of all non-Dagger Systems used for SWMG

Gravity now takes into account player's vertical velocity and updates it every tick, so which means that gravitational pull should feel more natural now.
