+++ include/GameSpecificSystems.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "Platform.h"
+#include "Gravity.h"
+#include "Pickup.h"
+#include "PlayerControls.h"
+#include "Combat.h"
