+++ include/PlayerControls.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Definitions.h"
+#include "Access.h"
+#include "GameSpecificComponents.h"
+
+struct PlayerControlsSystem 
+	: public ecs::System
+	, public MutAccessGroupStorage<Player, KeyBindings, Position, SpriteAnimation, Flip>
+{
+	void on_tick();
+};
